+++ python/marvin/tools/modelcube.py
@@ -6,7 +6,7 @@
 # @Author: Brian Cherinka
 # @Date:   2016-09-15 14:50:00
 # @Last modified by:   Brian Cherinka
-# @Last Modified time: 2016-10-11 20:01:40
+# @Last Modified time: 2017-02-05 00:07:24
 
 from __future__ import print_function, division, absolute_import